requirements.txt
numpy >= 1.14.0
PyAudio >= 0.2.11
pygame >= 1.9.0
ruamel.yaml >= 0.15.37
scipy >= 1.0.0
